Refractored the package to include cross encoder
